refactor: move class/functions into Gestures.hpp for reuse
