+++ src/gestures/Gestures.hpp
@@ -40,6 +40,30 @@ enum eTouchGestureDirection {
     // GESTURE_DIRECTION_OUT = (1 << 5),
 };
 
+// can be one of @eTouchGestureDirection or a combination of them
+using gestureDirection = uint32_t;
+
+// swipe and with multiple fingers and directions
+class CMultiAction : public wf::touch::gesture_action_t {
+  public:
+    CMultiAction(double threshold) : threshold(threshold){};
+    // bool pinch;
+    // bool last_pinch_was_pinch_in = false;
+    double threshold;
+
+    gestureDirection target_direction = 0;
+    int finger_count                  = 0;
+
+    wf::touch::action_status_t
+    update_state(const wf::touch::gesture_state_t& state,
+                 const wf::touch::gesture_event_t& event) override;
+
+    void reset(uint32_t time) override {
+        gesture_action_t::reset(time);
+        target_direction = 0;
+    };
+};
+
 std::unique_ptr<wf::touch::gesture_t>
 newWorkspaceSwipeStartGesture(const double sensitivity, const int fingers,
                               wf::touch::gesture_callback_t completed_cb,
